+++ sdk/core/azure-core/test/ut/modified_conditions.cpp
@@ -7,7 +7,6 @@
 #include <string>
 #include <vector>
 
-using namespace Azure::Core;
 using namespace Azure;
 
 TEST(ModifiedConditions, Basic)