+++ sdk/core/azure-core/test/ut/nullable.cpp
@@ -6,7 +6,7 @@
 #include <string>
 #include <vector>
 
-using namespace Azure::Core;
+using namespace Azure;
 
 TEST(Nullable, Basic)
 {