+++ sdk/core/azure-core/test/ut/simplified_header.cpp
@@ -40,7 +40,7 @@ TEST(SimplifiedHeader, core)
   EXPECT_NO_THROW(Azure::Core::Diagnostics::Logger::Listener ll = nullptr);
   EXPECT_NO_THROW(Azure::MatchConditions mc);
   EXPECT_NO_THROW(Azure::ModifiedConditions mc);
-  EXPECT_NO_THROW(Azure::Core::Nullable<int> n);
+  EXPECT_NO_THROW(Azure::Nullable<int> n);
   EXPECT_NO_THROW(Azure::Core::Uuid::CreateUuid());
   EXPECT_NO_THROW(Azure::Core::RequestFailedException("foo"));
   EXPECT_NO_THROW(Azure::Core::OperationStatus("foo"));