+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_options.hpp
@@ -34,7 +34,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief Filters the results to return only entries whose name begins with the specified
      * prefix.
      */
-    Azure::Core::Nullable<std::string> Prefix;
+    Azure::Nullable<std::string> Prefix;
 
     /**
      * @brief A string value that identifies the portion of the list to be returned with the next
@@ -42,19 +42,19 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * returned was not complete. The marker value may then be used in a subsequent call to request
      * the next set of list items. The marker value is opaque to the client.
      */
-    Azure::Core::Nullable<std::string> ContinuationToken;
+    Azure::Nullable<std::string> ContinuationToken;
 
     /**
      * @brief Specifies the maximum number of entries to return. If the request does not specify
      * PageSizeHint, or specifies a value greater than 5,000, the server will return up to 5,000
      * items.
      */
-    Azure::Core::Nullable<int32_t> PageSizeHint;
+    Azure::Nullable<int32_t> PageSizeHint;
 
     /**
      * @brief Include this parameter to specify one or more datasets to include in the response.
      */
-    Azure::Core::Nullable<Models::ListSharesIncludeType> ListSharesIncludeFlags;
+    Azure::Nullable<Models::ListSharesIncludeType> ListSharesIncludeFlags;
   };
 
   struct SetServicePropertiesOptions
@@ -76,12 +76,12 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief Specifies the access tier of the share. This is only valid for standard file account
      * and the value can only be one of `Hot`, `Cool` or `TransactionOptimized`
      */
-    Azure::Core::Nullable<Models::ShareAccessTier> AccessTier;
+    Azure::Nullable<Models::ShareAccessTier> AccessTier;
 
     /**
      * @brief Specifies the maximum size of the share, in gigabytes.
      */
-    Azure::Core::Nullable<int64_t> ShareQuotaInGiB;
+    Azure::Nullable<int64_t> ShareQuotaInGiB;
   };
 
   struct DeleteShareOptions
@@ -89,7 +89,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     /**
      * @brief Specifies the option include to delete the base share and all of its snapshots.
      */
-    Azure::Core::Nullable<bool> DeleteSnapshots;
+    Azure::Nullable<bool> DeleteSnapshots;
   };
 
   struct CreateShareSnapshotOptions
@@ -110,12 +110,12 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief Specifies the access tier of the share. This is only valid for standard file account
      * and the value can only be one of `Hot`, `Cool` or `TransactionOptimized`
      */
-    Azure::Core::Nullable<Models::ShareAccessTier> AccessTier;
+    Azure::Nullable<Models::ShareAccessTier> AccessTier;
 
     /**
      * @brief Specifies the maximum size of the share, in gigabytes.
      */
-    Azure::Core::Nullable<int64_t> ShareQuotaInGiB;
+    Azure::Nullable<int64_t> ShareQuotaInGiB;
   };
 
   struct SetShareMetadataOptions
@@ -175,7 +175,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * available before the break period has expired, but the lease may be held for longer than the
      * break period.
      */
-    Azure::Core::Nullable<int32_t> BreakPeriod;
+    Azure::Nullable<int32_t> BreakPeriod;
   };
 
   /**
@@ -196,7 +196,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief This permission is the security descriptor for the directory specified in the Security
      * Descriptor Definition Language (SDDL). If not specified, 'inherit' is used.
      */
-    Azure::Core::Nullable<std::string> DirectoryPermission;
+    Azure::Nullable<std::string> DirectoryPermission;
 
     /**
      * @brief SMB properties to set for the directory.
@@ -220,7 +220,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * shall be used. Default value: 'inherit'. If SDDL is specified as input, it must have owner,
      * group and dacl.
      */
-    Azure::Core::Nullable<std::string> FilePermission;
+    Azure::Nullable<std::string> FilePermission;
   };
 
   struct SetShareDirectoryMetadataOptions
@@ -233,7 +233,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief Filters the results to return only entries whose name begins with the specified
      * prefix.
      */
-    Azure::Core::Nullable<std::string> Prefix;
+    Azure::Nullable<std::string> Prefix;
 
     /**
      * @brief A string value that identifies the portion of the list to be returned with the next
@@ -241,14 +241,14 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * returned was not complete. The marker value may then be used in a subsequent call to request
      * the next set of list items. The marker value is opaque to the client.
      */
-    Azure::Core::Nullable<std::string> ContinuationToken;
+    Azure::Nullable<std::string> ContinuationToken;
 
     /**
      * @brief Specifies the maximum number of entries to return. If the request does not specify
      * PageSizeHint, or specifies a value greater than 5,000, the server will return up to 5,000
      * items.
      */
-    Azure::Core::Nullable<int32_t> PageSizeHint;
+    Azure::Nullable<int32_t> PageSizeHint;
   };
 
   struct ListShareDirectoryHandlesSinglePageOptions
@@ -259,20 +259,20 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * returned was not complete. The marker value may then be used in a subsequent call to request
      * the next set of list items. The marker value is opaque to the client.
      */
-    Azure::Core::Nullable<std::string> ContinuationToken;
+    Azure::Nullable<std::string> ContinuationToken;
 
     /**
      * @brief Specifies the maximum number of entries to return. If the request does not specify
      * PageSizeHint, or specifies a value greater than 5,000, the server will return up to 5,000
      * items.
      */
-    Azure::Core::Nullable<int32_t> PageSizeHint;
+    Azure::Nullable<int32_t> PageSizeHint;
 
     /**
      * @brief Specifies operation should apply to the directory specified in the URI, its files, its
      * subdirectories and their files.
      */
-    Azure::Core::Nullable<bool> Recursive;
+    Azure::Nullable<bool> Recursive;
   };
 
   struct ForceCloseShareDirectoryHandleOptions
@@ -287,13 +287,13 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * close was not complete. The marker value may then be used in a subsequent call to
      * close the next handle. The marker value is opaque to the client.
      */
-    Azure::Core::Nullable<std::string> ContinuationToken;
+    Azure::Nullable<std::string> ContinuationToken;
 
     /**
      * @brief Specifies operation should apply to the directory specified in the URI, its files, its
      * subdirectories and their files.
      */
-    Azure::Core::Nullable<bool> Recursive;
+    Azure::Nullable<bool> Recursive;
   };
 
   struct CreateShareFileOptions
@@ -302,7 +302,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief This permission is the security descriptor for the file specified in the Security
      * Descriptor Definition Language (SDDL). If not specified, 'inherit' is used.
      */
-    Azure::Core::Nullable<std::string> Permission;
+    Azure::Nullable<std::string> Permission;
 
     /**
      * @brief SMB properties to set for the file.
@@ -338,13 +338,13 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     /**
      * @brief Downloads only the bytes of the file from this range.
      */
-    Azure::Core::Nullable<Core::Http::HttpRange> Range;
+    Azure::Nullable<Core::Http::HttpRange> Range;
 
     /**
      * @brief When specified together with Range, service returns hash for the range as long as the
      * range is less than or equal to 4 MiB in size. Only MD5 is supported for now.
      */
-    Azure::Core::Nullable<HashAlgorithm> RangeHashAlgorithm;
+    Azure::Nullable<HashAlgorithm> RangeHashAlgorithm;
 
     /**
      * @brief The operation will only succeed if the access condition is met.
@@ -363,7 +363,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief This permission is the security descriptor for the file specified in the Security
      * Descriptor Definition Language (SDDL). If not specified, 'inherit' is used.
      */
-    Azure::Core::Nullable<std::string> Permission;
+    Azure::Nullable<std::string> Permission;
 
     /**
      * @brief SMB properties to set for the destination file.
@@ -374,19 +374,19 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief Specifies the option to copy file security descriptor from source file or to set it
      * using the value which is defined by the smb properties.
      */
-    Azure::Core::Nullable<Models::PermissionCopyModeType> PermissionCopyMode;
+    Azure::Nullable<Models::PermissionCopyModeType> PermissionCopyMode;
 
     /**
      * @brief Specifies the option to overwrite the target file if it already exists and has
      * read-only attribute set.
      */
-    Azure::Core::Nullable<bool> IgnoreReadOnly;
+    Azure::Nullable<bool> IgnoreReadOnly;
 
     /**
      * @brief Specifies the option to set archive attribute on a target file. True means archive
      * attribute will be set on a target file despite attribute overrides or a source file state.
      */
-    Azure::Core::Nullable<bool> SetArchiveAttribute;
+    Azure::Nullable<bool> SetArchiveAttribute;
 
     /**
      * @brief The operation will only succeed if the access condition is met.
@@ -416,12 +416,12 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief This permission is the security descriptor for the file specified in the Security
      * Descriptor Definition Language (SDDL). If not specified, 'inherit' is used.
      */
-    Azure::Core::Nullable<std::string> Permission;
+    Azure::Nullable<std::string> Permission;
 
     /**
      * @brief Specify this to resize a file to the specified value.
      */
-    Azure::Core::Nullable<int64_t> Size;
+    Azure::Nullable<int64_t> Size;
 
     /**
      * @brief The operation will only succeed if the access condition is met.
@@ -445,7 +445,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * compares the hash of the content that has arrived with the header value that was sent. If the
      * two hashes do not match, the operation will fail with error code 400 (Bad Request).
      */
-    Azure::Core::Nullable<ContentHash> TransactionalContentHash;
+    Azure::Nullable<ContentHash> TransactionalContentHash;
 
     /**
      * @brief The operation will only succeed if the access condition is met.
@@ -467,7 +467,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * @brief Specify the hash calculated for the range of bytes that must be read from the copy
      * source.
      */
-    Azure::Core::Nullable<ContentHash> TransactionalContentHash;
+    Azure::Nullable<ContentHash> TransactionalContentHash;
 
     /**
      * @brief Specify the access condition for the source. Only ContentHash with Crc64 is supported.
@@ -485,13 +485,13 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     /**
      * @brief The range to be get from service.
      */
-    Azure::Core::Nullable<Core::Http::HttpRange> Range;
+    Azure::Nullable<Core::Http::HttpRange> Range;
 
     /**
      * @brief The previous snapshot parameter is an opaque DateTime value that, when present,
      * specifies the previous snapshot.
      */
-    Azure::Core::Nullable<std::string> PreviousShareSnapshot;
+    Azure::Nullable<std::string> PreviousShareSnapshot;
 
     /**
      * @brief The operation will only succeed if the access condition is met.
@@ -507,14 +507,14 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * returned was not complete. The marker value may then be used in a subsequent call to request
      * the next set of list items. The marker value is opaque to the client.
      */
-    Azure::Core::Nullable<std::string> ContinuationToken;
+    Azure::Nullable<std::string> ContinuationToken;
 
     /**
      * @brief Specifies the maximum number of entries to return. If the request does not specify
      * PageSizeHint, or specifies a value greater than 5,000, the server will return up to 5,000
      * items.
      */
-    Azure::Core::Nullable<int32_t> PageSizeHint;
+    Azure::Nullable<int32_t> PageSizeHint;
   };
 
   struct ForceCloseShareFileHandleOptions
@@ -529,7 +529,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * close was not complete. The marker value may then be used in a subsequent call to
      * close the next handle. The marker value is opaque to the client.
      */
-    Azure::Core::Nullable<std::string> ContinuationToken;
+    Azure::Nullable<std::string> ContinuationToken;
   };
 
   /**
@@ -540,7 +540,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
     /**
      * @brief Downloads only the bytes of the file from this range.
      */
-    Azure::Core::Nullable<Core::Http::HttpRange> Range;
+    Azure::Nullable<Core::Http::HttpRange> Range;
 
     struct
     {
@@ -589,7 +589,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
      * shall be used. Default value: 'inherit'. If SDDL is specified as input, it must have owner,
      * group and dacl.
      */
-    Azure::Core::Nullable<std::string> FilePermission;
+    Azure::Nullable<std::string> FilePermission;
 
     struct
     {