+++ sdk/storage/azure-storage-files-shares/inc/azure/storage/files/shares/share_sas_builder.hpp
@@ -137,7 +137,7 @@ namespace Azure { namespace Storage { namespace Sas {
      * @brief Optionally specify the time at which the shared access signature becomes
      * valid. This timestamp will be truncated to second.
      */
-    Azure::Core::Nullable<Azure::DateTime> StartsOn;
+    Azure::Nullable<Azure::DateTime> StartsOn;
 
     /**
      * @brief The time at which the shared access signature becomes invalid. This field must
@@ -152,7 +152,7 @@ namespace Azure { namespace Storage { namespace Sas {
      * or address range specified on the SAS token, the request is not authenticated. When
      * specifying a range of IP addresses, note that the range is inclusive.
      */
-    Azure::Core::Nullable<std::string> IPRange;
+    Azure::Nullable<std::string> IPRange;
 
     /**
      * @brief An optional unique value up to 64 characters in length that correlates to an